+++ src/config.ts
@@ -1,11 +1,12 @@
 export default {
   //https://docs.ens.domains/ens-deployments
-  deployments: {
-    mainnet: '0x314159265dd8dbb310642f98f50c066173c1259b',
-    ropsten: '0x112234455c3a32fd11230c42e7bccd4a84e02010',
-    rinkeby: '0xe7410170f87102df0055eb195163a03b7f2bff4a',
-    goerly: '0x112234455c3a32fd11230c42e7bccd4a84e02010'
+  networkContract: {
+    1: '0x314159265dd8dbb310642f98f50c066173c1259b', // Mainnet
+    3: '0x112234455c3a32fd11230c42e7bccd4a84e02010', // Ropsten
+    4: '0xe7410170f87102df0055eb195163a03b7f2bff4a', // Rinkeby
+    5: '0x112234455c3a32fd11230c42e7bccd4a84e02010' // Goerli
   },
-  defaultNetworkName: 'mainnet',
-  scan_url: 'https://etherscan.io/address'
+  // defaultNetworkName: 'mainnet',
+  scan_url: 'https://etherscan.io/address',
+  networkURL: 'https://eth.gateway.whoisens.org'
 }