+++ src/index.ts
@@ -1,9 +1,19 @@
 import config from './config.js';
 
-import {EthAddressType, ISetResponse, ISetResponseError, IResponse, ResolveType} from './lib/types.js';
+import {
+    EthAddressType, ISetResponse, ISetResponseError, IResponse, ResolveType,
+    ResolverNotSetError, InvalidEthNameError, NameIsNotRegisteredError
+} from './lib/types.js';
+
 import ContractNames from './lib/ContractNames.js';
 import ENS from './lib/ENS.js';
 import utils from './utils/index.js';
+
+const Errors = {
+    ResolverNotSetError,
+    InvalidEthNameError,
+    NameIsNotRegisteredError
+};
 
 export {
     ResolveType,
@@ -13,7 +23,8 @@ export {
     ContractNames,
     EthAddressType,
     utils,
-    config
+    config,
+    Errors
 };
 
 export default ENS;