+++ src/lib/Config.ts
@@ -1,8 +1,9 @@
 export default class Config {
     private static instance: Config;
 
-    private currentNetwork: string;
-    private currentNetworkURL: string;
+    public currentNetworkId: string;
+    public currentNetworkURL: string;
+    public contractAddress: string;
 
     private constructor() {
     }
@@ -10,21 +11,5 @@ export default class Config {
     static getInstance() {
         if (!Config.instance) Config.instance = new Config();
         return Config.instance;
-    }
-
-    public getCurrentNetwork(): string {
-        return this.currentNetwork;
-    }
-
-    public getCurrentNetworkURL(): string {
-        return this.currentNetworkURL;
-    }
-
-    public setCurrentNetwork(currentNetwork: string): void {
-        this.currentNetwork = currentNetwork;
-    }
-
-    public setCurrentNetworkURL(currentNetworkURL: string): void {
-        this.currentNetworkURL = currentNetworkURL;
     }
 }