+++ src/lib/ENSRoot.ts
@@ -1,31 +1,27 @@
 import config from './../config.js';
-
 import utils from '../utils/index.js';
-import jsonRCP from '../utils/json-rcp.js';
+import jsonRCP from '../utils/json-rpc.js';
 
 import BaseClass from './BaseClass.js';
 import Config from './Config.js';
 import {IResponseResponseInfo} from './types.js';
 
 export default class ENSRoot extends BaseClass {
-    private readonly contractAddress: string;
-
     constructor() {
         super();
-
-        this.contractAddress = this.getContractAddress();
     }
 
     public async getController(address): Promise<IResponseResponseInfo> {
+        await this.init();
+
         const method = 'owner(bytes32)';
         const methodId = utils.getMethodID(method);
 
         const addressNode = utils.remove0x(utils.node(address));
 
-        const data = [methodId, addressNode].join('');
+        const data = `${methodId}${addressNode}`;
 
-        const result = await jsonRCP.makeRequest({
-            url: Config.getInstance().getCurrentNetworkURL(),
+        const result = await jsonRCP.getInstance().makeRequest({
             to: this.contractAddress,
             data
         });
@@ -39,11 +35,13 @@ export default class ENSRoot extends BaseClass {
                 addressNode
             },
             jsonRCPResult: result,
-            result: utils.normalizeHex(result.data.result)
+            result: utils.normalizeHex(result.result)
         });
     }
 
     public async getResolver(address): Promise<IResponseResponseInfo> {
+        await this.init();
+
         const method = 'resolver(bytes32)';
         const methodId = utils.getMethodID(method);
 
@@ -51,8 +49,7 @@ export default class ENSRoot extends BaseClass {
 
         const data = [methodId, addressNode].join('');
 
-        const result = await jsonRCP.makeRequest({
-            url: Config.getInstance().getCurrentNetworkURL(),
+        const result = await jsonRCP.getInstance().makeRequest({
             to: this.contractAddress,
             data
         });
@@ -66,11 +63,23 @@ export default class ENSRoot extends BaseClass {
                 addressNode
             },
             jsonRCPResult: result,
-            result: utils.normalizeHex(result.data.result)
+            result: utils.normalizeHex(result.result)
         });
     }
 
-    public getContractAddress(): string {
-        return config.deployments[Config.getInstance().getCurrentNetwork()];
+    public async findContractAddress(): Promise<string> {
+        const networkId = await jsonRCP.getInstance().getNetworkID();
+        Config.getInstance().currentNetworkId = networkId;
+
+        return config.networkContract[networkId];
+    }
+
+    private async init(): Promise<void> {
+        this.contractAddress = Config.getInstance().contractAddress;
+
+        if (!this.contractAddress) {
+            this.contractAddress = await this.findContractAddress();
+            Config.getInstance().contractAddress = this.contractAddress;
+        }
     }
 }